refactor(webui): move media replay helpers out of websocket channel
